consulgest-api/requirements.txt
django==4.0.3
django-environ
requests==2.27.1
djangorestframework==3.13.1
pandas==1.4.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-allauth==0.51.0
dj-rest-auth==2.2.5
djangorestframework-simplejwt==5.2.2
django-multiselectfield==0.1.12
skimpy~=0.0.7
pytest~=7.2.0
numpy~=1.24.0
keras~=2.10.0
tensorflow~=2.10.0
matplotlib~=3.6.2
pyspark~=3.2.1
scikit-learn~=1.1.3
imblearn~=0.0
jsonfield==3.1.0
psycopg2
django-postgres-extensions
django-filter==22.1
category-encoders==2.6.0
django-smtp-ssl==1.0